consensus: Rename contract_state.rs to operations.rs
