+++ consensus/src/proposal/step.rs
@@ -5,9 +5,9 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use crate::commons::{ConsensusError, Database};
-use crate::contract_state::Operations;
 use crate::execution_ctx::ExecutionCtx;
 use crate::msg_handler::{HandleMsgOutput, MsgHandler};
+use crate::operations::Operations;
 use node_data::message::Message;
 use std::cmp;
 use std::sync::Arc;
