+++ consensus/src/ratification/step.rs
@@ -5,8 +5,8 @@
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
 use crate::commons::{ConsensusError, Database, RoundUpdate};
-use crate::contract_state::Operations;
 use crate::execution_ctx::ExecutionCtx;
+use crate::operations::Operations;
 use std::marker::PhantomData;
 
 use crate::msg_handler::{HandleMsgOutput, MsgHandler};
